requirements.txt
Flask==3.0.3
Flask-Sessions==0.1.5
openai==1.30.1
python-dotenv==1.0.1
